+++ compatibility/features/attachments/attachments.ts
@@ -86,3 +86,26 @@ When('the {word} png is attached', async function (filename) {
     'image/png'
   )
 })
+
+When(
+  'a PDF document is attached with a filename',
+  async function (this: World) {
+    await this.attach(
+      fs.createReadStream(
+        path.join(
+          process.cwd(),
+          'node_modules',
+          '@cucumber',
+          'compatibility-kit',
+          'features',
+          'attachments',
+          'document.pdf'
+        )
+      ),
+      {
+        mediaType: 'application/pdf',
+        fileName: 'document.pdf',
+      }
+    )
+  }
+)