+++ src/api/support.ts
@@ -2,6 +2,7 @@ import { IdGenerator } from '@cucumber/messages'
 import { ISupportCodeLibrary } from '../support_code_library_builder/types'
 import supportCodeLibraryBuilder from '../support_code_library_builder'
 import { pathToFileURL } from 'url'
+import tryRequire from '../try_require'
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { importer } = require('../importer')
@@ -24,10 +25,13 @@ export async function getSupportCodeLibrary({
     requirePaths,
     importPaths,
   })
-  requireModules.map((module) => require(module))
-  requirePaths.map((path) => require(path))
+
+  requireModules.map((module) => tryRequire(module))
+  requirePaths.map((path) => tryRequire(path))
+
   for (const path of importPaths) {
     await importer(pathToFileURL(path))
   }
+
   return supportCodeLibraryBuilder.finalize()
 }