+++ src/cli/index.ts
@@ -1,4 +1,4 @@
-import { ArgvParser, isTruthyString } from '../configuration'
+import { ArgvParser } from '../configuration'
 import { IFormatterStream } from '../formatter'
 import { loadConfiguration, runCucumber } from '../api'
 import { getKeywords, getLanguages } from './i18n'
@@ -6,7 +6,6 @@ import { validateInstall } from './install_validator'
 import debug from 'debug'
 
 export interface ICliRunResult {
-  shouldAdvertisePublish: boolean
   shouldExitImmediately: boolean
   success: boolean
 }
@@ -39,14 +38,16 @@ export default class Cli {
   }
 
   async run(): Promise<ICliRunResult> {
-    await validateInstall()
+    const debugEnabled = debug.enabled('cucumber')
+    if (debugEnabled) {
+      await validateInstall()
+    }
     const { options, configuration: argvConfiguration } = ArgvParser.parse(
       this.argv
     )
     if (options.i18nLanguages) {
       this.stdout.write(getLanguages())
       return {
-        shouldAdvertisePublish: false,
         shouldExitImmediately: true,
         success: true,
       }
@@ -54,17 +55,17 @@ export default class Cli {
     if (options.i18nKeywords) {
       this.stdout.write(getKeywords(options.i18nKeywords))
       return {
-        shouldAdvertisePublish: false,
         shouldExitImmediately: true,
         success: true,
       }
     }
+
     const environment = {
       cwd: this.cwd,
       stdout: this.stdout,
       stderr: this.stderr,
       env: this.env,
-      debug: debug.enabled('cucumber'),
+      debug: debugEnabled,
     }
     const { useConfiguration: configuration, runConfiguration } =
       await loadConfiguration(
@@ -77,10 +78,6 @@ export default class Cli {
       )
     const { success } = await runCucumber(runConfiguration, environment)
     return {
-      shouldAdvertisePublish:
-        !runConfiguration.formats.publish &&
-        !configuration.publishQuiet &&
-        !isTruthyString(this.env.CUCUMBER_PUBLISH_QUIET),
       shouldExitImmediately: configuration.forceExit,
       success,
     }