+++ src/cli/install_validator.ts
@@ -6,7 +6,7 @@ export async function validateInstall(): Promise<void> {
     console.warn(
       `
       It looks like you're running Cucumber from a global installation.
-      This won't work - you need to have Cucumber installed as a local dependency in your project.
+      If so, you'll likely see issues - you need to have Cucumber installed as a local dependency in your project.
       See https://github.com/cucumber/cucumber-js/blob/main/docs/installation.md#invalid-installations
       `
     )