+++ src/configuration/types.ts
@@ -1,12 +1,14 @@
 import { FormatOptions } from '../formatter'
 import { PickleOrder } from '../models/pickle_order'
 
+type FormatsConfiguration = Array<string | [string, string?]>
+
 export interface IConfiguration {
   backtrace: boolean
   dryRun: boolean
   forceExit: boolean
   failFast: boolean
-  format: string[]
+  format: FormatsConfiguration
   formatOptions: FormatOptions
   import: string[]
   language: string
@@ -15,6 +17,9 @@ export interface IConfiguration {
   paths: string[]
   parallel: number
   publish: boolean
+  /**
+   * @deprecated no longer needed; see <https://github.com/cucumber/cucumber-js/blob/main/docs/deprecations.md>
+   */
   publishQuiet: boolean
   require: string[]
   requireModule: string[]